Move db and params properties to core
This is partly so users can just require core and have access to
all the syntax necessary to execute statements. It's also because
these properties are universal to any execution back end. How you
specify a db or specific parameter values may differ between back
ends. But the fact of needing to specify a db and a collection of
parameters is independent of the backend.