Prep `ns` forms for CREATE by excluding conflicts
CREATE TRIGGER will introduce `before` and `after` functions
which conflict with the midje.sweet namespace. It also adds
a `when` function which conflicts with clojure.core namespace.

Since this commit affects a number of files, I made it first as
a preparatory step so that the important changes to add the
CREATE statement will be more apparent.