Move wiki content to `doc` folder as Asciidoc
This first commit just brings over the existing documentation.
I used the `kramdown` Asciidoctor extension to convert files,
but I had to make a few manual adjustments.

Subsequent commits will add, remove, or modify documentation.
This move is intended to bring the documentation in sync with
the code and prepare it for use with `cljdoc`.